Implemented a proper game loop
The computer currently only plays at (1,2,3). It always moves at this position. This is temporary.

Do "CTRL+F: TODO" inside the source file to find the TODO list.

Basically, I only need to implement minimax and heuristic analysis of the computer AI. Also, the getOutcome function to check if someone won the game.