Added functions which extract multiple 2D boards out of the single 3D board
Added:
- getUpDownBoards
- getLeftRightBoards
- getBackForwardBoards

TODO: I need to write a function which gets me the 4 diagonals out of the 3D board.

TODO: Inside the getOutcome function, when I'm finally checking the multiple "back-forward" 2D tictactoe boards, I need to make sure and send the TRANSPOSE of each of the 2D boards.

My brain is fried.